Generating the test files
This PR creates copies of the test file we have on disk to a git ignored folder. While
copying, it is also changing the file name to include special characters. The code also
takes care of created the folder and replacing its content as needed.

To test
Run PHP script as php testTree.php.
A new folder named generatedFiles will be created.
New files with filesnames containing special characters should be added to that folder.